+++ src/main/java/com/nhnacademy/member_server/controller/swagger/PointAdminApi.java
@@ -1,9 +1,9 @@
 package com.nhnacademy.member_server.controller.swagger;
 
-import com.nhnacademy.member_server.dto.request.PointAdminAdjustmentRequest;
-import com.nhnacademy.member_server.dto.request.PointAdminPolicyRequest;
-import com.nhnacademy.member_server.dto.response.PointAdminPolicyResponse;
-import com.nhnacademy.member_server.dto.response.PointTransactionResponse;
+import com.nhnacademy.member_server.dto.request.point.PointAdminAdjustmentRequest;
+import com.nhnacademy.member_server.dto.request.point.PointAdminPolicyRequest;
+import com.nhnacademy.member_server.dto.response.point.PointAdminPolicyResponse;
+import com.nhnacademy.member_server.dto.response.point.PointTransactionResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.parameters.RequestBody;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
