+++ src/main/java/com/nhnacademy/member_server/dto/response/cart/CartAddResponse.java
@@ -1,3 +1,3 @@
-package com.nhnacademy.member_server.dto.cartResponse;
+package com.nhnacademy.member_server.dto.response.cart;
 
 public record CartAddResponse(String key, Long bookId, int quantity) {}
