+++ src/main/java/com/nhnacademy/member_server/dto/response/cart/GetBookResponse.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.member_server.dto.cartResponse;
+package com.nhnacademy.member_server.dto.response.cart;
 
 public record GetBookResponse(Long bookId,
                               String title,
