+++ src/main/java/com/nhnacademy/member_server/repository/CartItemRepository.java
@@ -1,6 +1,6 @@
 package com.nhnacademy.member_server.repository;
 
-import com.nhnacademy.member_server.entity.cartEntity.CartItem;
+import com.nhnacademy.member_server.entity.cart.CartItem;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
