+++ src/main/java/com/nhnacademy/member_server/repository/CartRepository.java
@@ -1,6 +1,6 @@
 package com.nhnacademy.member_server.repository;
 
-import com.nhnacademy.member_server.entity.cartEntity.Cart;
+import com.nhnacademy.member_server.entity.cart.Cart;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
