+++ src/main/java/com/nhnacademy/member_server/repository/PointPolicyRepository.java
@@ -1,6 +1,6 @@
 package com.nhnacademy.member_server.repository;
 
-import com.nhnacademy.member_server.entity.PointPolicy;
+import com.nhnacademy.member_server.entity.point.PointPolicy;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PointPolicyRepository extends JpaRepository<PointPolicy, Long> {
