+++ src/main/java/com/nhnacademy/member_server/service/CartService.java
@@ -1,10 +1,10 @@
 package com.nhnacademy.member_server.service;
 
-import com.nhnacademy.member_server.dto.cartRequest.CartAddRequest;
-import com.nhnacademy.member_server.dto.cartRequest.CartItemUpdateRequest;
-import com.nhnacademy.member_server.dto.cartResponse.CartAddResponse;
-import com.nhnacademy.member_server.dto.cartResponse.CartListResponse;
-import com.nhnacademy.member_server.dto.cartResponse.CartUpdateResponse;
+import com.nhnacademy.member_server.dto.request.cart.CartAddRequest;
+import com.nhnacademy.member_server.dto.request.cart.CartItemUpdateRequest;
+import com.nhnacademy.member_server.dto.response.cart.CartAddResponse;
+import com.nhnacademy.member_server.dto.response.cart.CartListResponse;
+import com.nhnacademy.member_server.dto.response.cart.CartUpdateResponse;
 import java.util.Map;
 
 public interface CartService {
