+++ src/main/java/com/nhnacademy/member_server/service/impl/PointServiceImpl.java
@@ -6,17 +6,17 @@ import static com.nhnacademy.member_server.exception.ErrorCode.POINT_NOT_ENOUGH;
 import static com.nhnacademy.member_server.exception.ErrorCode.POINT_NOT_ORDER_ID;
 import static com.nhnacademy.member_server.exception.ErrorCode.POINT_POLICY_NOT_FOUND;
 
-import com.nhnacademy.member_server.dto.request.PointAdminAdjustmentRequest;
-import com.nhnacademy.member_server.dto.request.PointAdminPolicyRequest;
-import com.nhnacademy.member_server.dto.request.PointTransactionCreateRequest;
-import com.nhnacademy.member_server.dto.response.PointAdminPolicyResponse;
-import com.nhnacademy.member_server.dto.response.PointBalanceResponse;
-import com.nhnacademy.member_server.dto.response.PointHistoryResponse;
-import com.nhnacademy.member_server.entity.PointEventType;
-import com.nhnacademy.member_server.entity.PointHistory;
-import com.nhnacademy.member_server.entity.PointPolicy;
-import com.nhnacademy.member_server.entity.PointStatus;
+import com.nhnacademy.member_server.dto.request.point.PointAdminAdjustmentRequest;
+import com.nhnacademy.member_server.dto.request.point.PointAdminPolicyRequest;
+import com.nhnacademy.member_server.dto.request.point.PointTransactionCreateRequest;
+import com.nhnacademy.member_server.dto.response.point.PointAdminPolicyResponse;
+import com.nhnacademy.member_server.dto.response.point.PointBalanceResponse;
+import com.nhnacademy.member_server.dto.response.point.PointHistoryResponse;
 import com.nhnacademy.member_server.entity.member.Member;
+import com.nhnacademy.member_server.entity.point.PointEventType;
+import com.nhnacademy.member_server.entity.point.PointHistory;
+import com.nhnacademy.member_server.entity.point.PointPolicy;
+import com.nhnacademy.member_server.entity.point.PointStatus;
 import com.nhnacademy.member_server.exception.BusinessException;
 import com.nhnacademy.member_server.exception.ErrorCode;
 import com.nhnacademy.member_server.repository.MemberRepository;
