+++ src/test/java/com/nhnacademy/member_server/service/CartServiceTest.java
@@ -15,14 +15,14 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 
-import com.nhnacademy.member_server.dto.cartRequest.CartAddRequest;
-import com.nhnacademy.member_server.dto.cartRequest.CartItemUpdateRequest;
-import com.nhnacademy.member_server.dto.cartResponse.CartAddResponse;
-import com.nhnacademy.member_server.dto.cartResponse.CartListResponse;
-import com.nhnacademy.member_server.dto.cartResponse.CartUpdateResponse;
-import com.nhnacademy.member_server.dto.cartResponse.GetBookResponse;
-import com.nhnacademy.member_server.entity.cartEntity.Cart;
-import com.nhnacademy.member_server.entity.cartEntity.CartItem;
+import com.nhnacademy.member_server.dto.request.cart.CartAddRequest;
+import com.nhnacademy.member_server.dto.request.cart.CartItemUpdateRequest;
+import com.nhnacademy.member_server.dto.response.cart.CartAddResponse;
+import com.nhnacademy.member_server.dto.response.cart.CartListResponse;
+import com.nhnacademy.member_server.dto.response.cart.CartUpdateResponse;
+import com.nhnacademy.member_server.dto.response.cart.GetBookResponse;
+import com.nhnacademy.member_server.entity.cart.Cart;
+import com.nhnacademy.member_server.entity.cart.CartItem;
 import com.nhnacademy.member_server.entity.member.Member;
 import com.nhnacademy.member_server.exception.BusinessException;
 import com.nhnacademy.member_server.exception.ErrorCode;
