+++ src/test/java/com/nhnacademy/member_server/service/PointConcurrencyTest.java
@@ -2,12 +2,12 @@ package com.nhnacademy.member_server.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.nhnacademy.member_server.dto.request.PointTransactionCreateRequest;
-import com.nhnacademy.member_server.entity.PointEventType;
+import com.nhnacademy.member_server.dto.request.point.PointTransactionCreateRequest;
 import com.nhnacademy.member_server.entity.member.Grade;
 import com.nhnacademy.member_server.entity.member.Member;
 import com.nhnacademy.member_server.entity.member.Role;
 import com.nhnacademy.member_server.entity.member.Status;
+import com.nhnacademy.member_server.entity.point.PointEventType;
 import com.nhnacademy.member_server.repository.GradeRepository;
 import com.nhnacademy.member_server.repository.MemberRepository;
 import com.nhnacademy.member_server.service.impl.social.PaycoLoginStrategy;
