+++ src/test/java/com/nhnacademy/member_server/service/PointServiceImplTest.java
@@ -7,15 +7,15 @@ import static org.mockito.BDDMockito.given;
 import static org.mockito.BDDMockito.then;
 import static org.mockito.Mockito.never;
 
-import com.nhnacademy.member_server.dto.request.PointAdminAdjustmentRequest;
-import com.nhnacademy.member_server.dto.request.PointAdminPolicyRequest;
-import com.nhnacademy.member_server.dto.request.PointTransactionCreateRequest;
-import com.nhnacademy.member_server.entity.PointEventType;
-import com.nhnacademy.member_server.entity.PointHistory;
-import com.nhnacademy.member_server.entity.PointPolicy;
-import com.nhnacademy.member_server.entity.PointStatus;
+import com.nhnacademy.member_server.dto.request.point.PointAdminAdjustmentRequest;
+import com.nhnacademy.member_server.dto.request.point.PointAdminPolicyRequest;
+import com.nhnacademy.member_server.dto.request.point.PointTransactionCreateRequest;
 import com.nhnacademy.member_server.entity.member.Grade;
 import com.nhnacademy.member_server.entity.member.Member;
+import com.nhnacademy.member_server.entity.point.PointEventType;
+import com.nhnacademy.member_server.entity.point.PointHistory;
+import com.nhnacademy.member_server.entity.point.PointPolicy;
+import com.nhnacademy.member_server.entity.point.PointStatus;
 import com.nhnacademy.member_server.exception.BusinessException;
 import com.nhnacademy.member_server.exception.ErrorCode;
 import com.nhnacademy.member_server.repository.MemberRepository;
